change the vesion

+++ ASNET/src/ASNET.Physics.cpp
@@ -7,7 +7,7 @@ namespace ASNET {
 		
 		float Vector2::Distance(float x1, float y1, float x2, float y2) 
 		{
-			return sqrt((x1 - x2)*(x1 - x2) + (y1 - y2)*(y1 - y2));
+			return (float)sqrt((x1 - x2)*(x1 - x2) + (y1 - y2)*(y1 - y2));
 		}
 		
 		Vector3 ASNET::Physics::Vector3::operator+(const Vector3 vector3)
